Stash `Commit` messages while a rebalance is unfinished. Fail actor of the rebalance doesn't finish within a timeout.
Fixes #750

+++ tests/src/test/scala/akka/kafka/scaladsl/TransactionsSpec.scala
@@ -253,7 +253,7 @@ class TransactionsSpec extends SpecBase(kafkaPort = KafkaPorts.TransactionsSpec)
       val probeConsumer = valuesProbeConsumer(probeConsumerSettings(probeConsumerGroup), sinkTopic)
 
       probeConsumer
-        .request(elements)
+        .request(elements.toLong)
         .expectNextUnorderedN((1 to elements).map(_.toString))
 
       probeConsumer.cancel()
